+++ comm/src/main/scala/io/casperlabs/comm/transport/HostnameTrustManagerFactory.scala
@@ -1,14 +1,9 @@
 package io.casperlabs.comm.transport
 
-import java.net.Socket
 import java.security.KeyStore
-import java.security.cert.{CertificateException, X509Certificate}
-
-import io.casperlabs.crypto.codec.Base16
-import io.casperlabs.crypto.util.CertificateHelper
-import io.netty.handler.ssl.util.SimpleTrustManagerFactory
-import io.netty.util.internal.EmptyArrays
 import javax.net.ssl._
+import io.casperlabs.crypto.util.HostnameTrustManager
+import io.netty.handler.ssl.util.SimpleTrustManagerFactory
 
 class HostnameTrustManagerFactory private () extends SimpleTrustManagerFactory {
   def engineInit(keyStore: KeyStore): Unit                                 = {}
@@ -21,95 +16,4 @@ class HostnameTrustManagerFactory private () extends SimpleTrustManagerFactory {
 
 object HostnameTrustManagerFactory {
   val Instance: HostnameTrustManagerFactory = new HostnameTrustManagerFactory()
-}
-
-private class HostnameTrustManager extends X509ExtendedTrustManager {
-
-  private def reject[T](msg: String): T =
-    throw new CertificateException(msg)
-
-  def checkClientTrusted(
-      x509Certificates: Array[X509Certificate],
-      authType: String,
-      socket: Socket
-  ): Unit =
-    reject("Not allowed validation method")
-
-  def checkClientTrusted(
-      x509Certificates: Array[X509Certificate],
-      authType: String,
-      sslEngine: SSLEngine
-  ): Unit = {
-    Option(sslEngine.getHandshakeSession)
-      .getOrElse(reject("No handshake session"))
-
-    val cert = x509Certificates.head
-    val peerHost = CertificateHelper
-      .publicAddress(cert.getPublicKey)
-      .map(Base16.encode)
-      .getOrElse(reject(s"Certificate's public key has the wrong algorithm"))
-    checkIdentity(Some(peerHost), cert, "https")
-  }
-
-  def checkClientTrusted(x509Certificates: Array[X509Certificate], authType: String): Unit =
-    reject("Not allowed validation method")
-
-  def checkServerTrusted(
-      x509Certificates: Array[X509Certificate],
-      authType: String,
-      socket: Socket
-  ): Unit =
-    reject("Not allowed validation method")
-
-  def checkServerTrusted(
-      x509Certificates: Array[X509Certificate],
-      authType: String,
-      sslEngine: SSLEngine
-  ): Unit = {
-    val sslSession = Option(sslEngine.getHandshakeSession)
-      .getOrElse(reject("No handshake session"))
-
-    // check endpoint identity
-    Option(sslEngine.getSSLParameters.getEndpointIdentificationAlgorithm) match {
-      case Some(identityAlg) if identityAlg.nonEmpty =>
-        val cert     = x509Certificates.head
-        val peerHost = Option(sslSession.getPeerHost)
-
-        checkIdentity(peerHost, cert, identityAlg)
-
-        CertificateHelper
-          .publicAddress(cert.getPublicKey)
-          .map(Base16.encode)
-          .filter(_ == peerHost.getOrElse(""))
-          .getOrElse(reject(s"Certificate's public address doesn't match the hostname"))
-
-      case _ =>
-        reject("No endpoint identification algorithm")
-    }
-  }
-
-  def checkServerTrusted(x509Certificates: Array[X509Certificate], authType: String): Unit =
-    reject("Not allowed validation method")
-
-  def getAcceptedIssuers: Array[X509Certificate] = EmptyArrays.EMPTY_X509_CERTIFICATES
-
-  private def checkIdentity(
-      hostname: Option[String],
-      cert: X509Certificate,
-      algorithm: String
-  ): Unit = {
-    import sun.security.util.HostnameChecker
-    algorithm.toLowerCase match {
-      case "https" =>
-        val host = hostname
-          .filter(_.startsWith("["))
-          .filter(_.endsWith("]"))
-          .map(h => h.substring(1, h.length - 1))
-          .orElse(hostname)
-          .getOrElse("")
-        HostnameChecker.getInstance(HostnameChecker.TYPE_TLS).`match`(host, cert)
-      case _ =>
-        reject(s"Unknown identification algorithm: $algorithm")
-    }
-  }
 }
